Move missed photo flow and rename folder for clarity
