+++ src/Language/Locales/en/translation.json
@@ -287,7 +287,7 @@
             "missedDays": {
                 "missing": "Missing Treatment Log",
                 "missing_plural": "Missing Treatment Logs",
-                "missedPhoto": "Submit missed photo"
+                "missedPhoto": "Submit missed test"
             },
             "chatReminder": {
                 "body": "Did you know you can chat anonymously with other patients?",
@@ -921,6 +921,7 @@
         "noResponse": "Not yet completed"
     },
     "missedPhotoDetails":{
+        "topBarText": "Missed request",
         "future": "In the future please submit your test on the day it is requested so we can see your treatment progress.",
         "requestedOn": "Submitting for the request sent on"
     }
