+++ src/Patient/ReportingFlows/SymptomWarning.js
